NewsPortal/NewsPortal/requirements.txt
amqp==5.1.1
APScheduler==3.9.1.post1
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==38.0.4
defusedxml==0.7.1
dill==0.3.6
Django==4.1.3
django-allauth==0.51.0
django-filter==22.1
dnspython==2.2.1
eventlet==0.33.2
greenlet==2.0.1
idna==3.4
isort==5.11.4
kombu==5.2.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==2.6.0
prompt-toolkit==3.0.36
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
